// frontend/src/api/generated/models/saved-search-save-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * chainlink-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SavedSearchSaveJson
 */
export interface SavedSearchSaveJson {
    /**
     * 
     * @type {string}
     * @memberof SavedSearchSaveJson
     */
    collectionId: string;
    /**
     * 
     * @type {string}
     * @memberof SavedSearchSaveJson
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof SavedSearchSaveJson
     */
    query: string;
}

/**
 * Check if a given object implements the SavedSearchSaveJson interface.
 */
export function instanceOfSavedSearchSaveJson(value: object): value is SavedSearchSaveJson {
    if (!('collectionId' in value) || value['collectionId'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('query' in value) || value['query'] === undefined) return false;
    return true;
}

export function SavedSearchSaveJsonFromJSON(json: any): SavedSearchSaveJson {
    return SavedSearchSaveJsonFromJSONTyped(json, false);
}

export function SavedSearchSaveJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): SavedSearchSaveJson {
    if (json == null) {
        return json;
    }
    return {
        
        'collectionId': json['collectionId'],
        'name': json['name'],
        'query': json['query'],
    };
}

export function SavedSearchSaveJsonToJSON(json: any): SavedSearchSaveJson {
    return SavedSearchSaveJsonToJSONTyped(json, false);
}

export function SavedSearchSaveJsonToJSONTyped(value?: SavedSearchSaveJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'collectionId': value['collectionId'],
        'name': value['name'],
        'query': value['query'],
    };
}